requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
gradio
pydantic
astor